Adds check for files/folder before deletion

+++ Anytime_RRT_Sharp/helper/library_generator.m
@@ -8,8 +8,12 @@ baseHppPath = fullfile(productPath, "src");
 
 % delete old versions
 delete defineArrtsLibrary.*
-delete ArrtsLibraryData.xml
-rmdir ArrtsLibrary s
+if (isfile("ArrtsLibraryData.xml"))
+    delete ArrtsLibraryData.xml
+end
+if (isfolder("ArrtsLibrary"))
+    rmdir ArrtsLibrary s
+end
 
 % create library array
 libs = ["libArrtsParams.a",...
